Unify line tests under one file and test for line text
